feat: Update logger configuration to redact sensitive information
The code changes implement a `get_logger` function that creates and configures a logger named 'user_data'. The logger is set to log at the INFO level and uses a `RedactingFormatter` to redact sensitive information fields specified in the `PII_FIELDS` list. This update enhances the security of the logging process by preventing sensitive information from being logged.